LayoutTests/imported/w3c/web-platform-tests/css/cssom/HTMLLinkElement-disabled-002-expected.txt


FAIL HTMLLinkElement.disabled reflects the <link disabled> attribute, and behaves consistently, when the sheet is an alternate assert_equals: expected null but got Element node <link title="alt" rel="alternate stylesheet" href="data:t...
